refactor image utils into imageutils
